Added startup file for Crestle
This file downloads the latest fastai course data, and updates
anaconda and the fastai library to the latest available.

If you have modified any of the Jupyter Notebooks, it's
possible the git pull will fail. If this happens, simply use
git stash to stash your updates and run git pull again. If you
want to bring the changes back in, use git stash to do that.