Added comments and made training data only download if not there
Added comments to make the the code more understandable. Also
altered the training data download to be conditional upon whether
the data has already been downloaded. The goal here is to avoid
overwriting any changes to the training dataset that the user has
made. It forces the user to delete the training data folder
explicitly if they want to refresh from Github to avoid accidental
loss of work.